+++ ChangeLog.md
@@ -17,6 +17,8 @@ Behavior changes:
 
 Other enhancements:
 
+* On Windows before Windows 10, --color=never is the default on terminals that
+  can support ANSI color codes in output only by emulation
 * On Windows, recognise a 'mintty' (false) terminal as a terminal, by default
 * `stack build` issues a warning when `base` is explicitly listed in
   `extra-deps` of `stack.yaml`